Aggiornamento delle pagine delle statistiche
