+++ src/app/stats/economy/page.tsx
@@ -2,11 +2,11 @@ import Navbar from "@/components/layouts/Navbar";
 import SplitText from "@/components/reactbits/SplitText";
 import AnimatedContent from "@/components/reactbits/AnimatedContent";
 import DotGrid from "@/components/reactbits/DotGrid";
-import {InsubricaUrbanGraph} from "@/components/ui/stats/urban/InSuUrbanGraph";
-import {ComoUrbanGraph} from "@/components/ui/stats/urban/sub/ComoUrbanGraph";
-import {LeccoUrbanGraph} from "@/components/ui/stats/urban/sub/LeccoUrbanGraph";
-import {TicinoUrbanGraph} from "@/components/ui/stats/urban/sub/TicinoUrbanGraph";
-import {VareseUrbanGraph} from "@/components/ui/stats/urban/sub/VareseUrbanGraph";
+import {InsubricaEconomySectorsGraph} from "@/components/ui/stats/economy/InSuEconomySectorsGraph";
+import {ComoEconomyGraph} from "@/components/ui/stats/economy/sub/ComoEconomyGraph";
+import {TicinoEconomyGraph} from "@/components/ui/stats/economy/sub/TicinoEconomyGraph";
+import {VareseEconomyGraph} from "@/components/ui/stats/economy/sub/VareseEconomyGraph";
+import {LeccoEconomyGraph} from "@/components/ui/stats/economy/sub/LeccoEconomyGraph";
 
 export default function GladysPage() {
     return (
@@ -15,7 +15,7 @@ export default function GladysPage() {
             <div id={"head"} className={"fixed z-10 w-full h-118.75 flex flex-col items-center justify-center"}>
                 <div className={"relative h-full z-1 flex flex-col items-center justify-center"}>
                     <SplitText
-                        text="Statistiche"
+                        text="Economia"
                         className="text-[114px] font-black text-center select-none cursor-default mb-0 mt-18"
                         ease="elastic.out(1, 0.3)"
                         splitType="chars"
@@ -30,7 +30,7 @@ export default function GladysPage() {
                         className={"mb-10"}
                         delay={0.6}
                     >
-                        <p className={"max-w-100 text-center mt-0"}>Approfondisci le statistiche per l&#39;urbanizzazione</p>
+                        <p className={"max-w-100 text-center mt-0"}>Approfondisci le statistiche per l'economia dell'area Insubrica</p>
                     </AnimatedContent>
                 </div>
                 <div id={"home-page-head-background"} className={"absolute -z-100 w-full h-[50vh]"}>
@@ -56,23 +56,23 @@ function InsubricaCharts() {
                 {/* Titolo */}
                 <div>
                     <h1 className="text-5xl font-bold text-white">
-                        Grafici sull&#39;Urbanizzazione
+                        Grafici Economici dell&#39;area Insubrica
                     </h1>
                     <p className="text-gray-100 mt-2 mb-20">
-                        Questa pagina è stata realizzata per visualizzare importanti grafici in merito all&#39;urbanizzazione
-                        della regione Insubrica.
+                        In questa pagina avrai modo di analizzare e studiare molti dati attuali in merito all'economia nella regione Insubrica,
+                        aggiornati all&#39;anno <b>2025</b>.
                     </p>
                 </div>
 
                 {/* Grafico principale grande */}
-                <InsubricaUrbanGraph/>
+                <InsubricaEconomySectorsGraph/>
 
                 {/* Sezione grafici secondari */}
                 <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-                    <ComoUrbanGraph/>
-                    <LeccoUrbanGraph/>
-                    <TicinoUrbanGraph/>
-                    <VareseUrbanGraph/>
+                    <LeccoEconomyGraph/>
+                    <ComoEconomyGraph/>
+                    <VareseEconomyGraph/>
+                    <TicinoEconomyGraph/>
                 </div>
             </div>
         </div>
