+++ src/app/stats/page.tsx
@@ -2,15 +2,16 @@ import Navbar from "@/components/layouts/Navbar";
 import SplitText from "@/components/reactbits/SplitText";
 import AnimatedContent from "@/components/reactbits/AnimatedContent";
 import DotGrid from "@/components/reactbits/DotGrid";
+import Footer from "@/components/layouts/Footer";
 
-export default function GladysPage() {
+export default function StatistichePage() {
     return (
         <div>
-            <Navbar page={5}/>
+            <Navbar page={4}/>
             <div id={"head"} className={"fixed z-10 w-full h-118.75 flex flex-col items-center justify-center"}>
                 <div className={"relative h-full z-1 flex flex-col items-center justify-center"}>
                     <SplitText
-                        text="Meteo"
+                        text="Statistiche"
                         className="text-[114px] font-black text-center select-none cursor-default mb-0 mt-18"
                         ease="elastic.out(1, 0.3)"
                         splitType="chars"
@@ -25,7 +26,7 @@ export default function GladysPage() {
                         className={"mb-10"}
                         delay={0.6}
                     >
-                        <p className={"max-w-100 text-center mt-0"}>Una soluzione metereologica/statistica innovativa</p>
+                        <p className={"max-w-100 text-center mt-0"}>Esplora le statistiche dell'area Insubrica</p>
                     </AnimatedContent>
                 </div>
                 <div id={"home-page-head-background"} className={"absolute -z-100 w-full h-[50vh]"}>
@@ -37,8 +38,13 @@ export default function GladysPage() {
                     />
                 </div>
             </div>
-            <div id={"page-content"} className={"relative z-20 w-full bg-[#0a0a0a] top-118.75 py-16"}>
+            <div id={"page-content"} className={"relative z-20 w-full bg-[#0a0a0a] top-118.75"}>
+                <div id={"content"} className={"w-full py-24 px-16 flex flex-col items-center justify-center"}>
+                    <div className="w-full max-w-375 flex flex-col">
 
+                    </div>
+                </div>
+                <Footer/>
             </div>
         </div>
     );
